from int to long

+++ src/dataAccess/UpdateFailedException.java
@@ -1,7 +1,7 @@
 package dataAccess;
 
 public class UpdateFailedException extends DAOOperationFailedException {
-    public UpdateFailedException(String objectType, int objectId, String message) {
+    public UpdateFailedException(String objectType, long objectId, String message) {
         super("update", objectType, objectId, message);
     }
 }
